+++ libs/components/docs-tools/src/lib/modules/code-example-viewer/stackblitz.service.ts
@@ -77,7 +77,7 @@ export class SkyDocsStackBlitzService {
             "outputPath": "dist/example-app",
             "index": "src/index.html",
             "main": "src/main.ts",
-            "polyfills": ["zone.js", "@skyux/packages/polyfills"],
+            "polyfills": ["zone.js"],
             "tsConfig": "tsconfig.app.json",
             "inlineStyleLanguage": "scss",
             "assets": ["src/assets"],
@@ -149,8 +149,7 @@ export class SkyDocsStackBlitzService {
             "main": "src/test.ts",
             "polyfills": [
               "zone.js",
-              "zone.js/testing",
-              "@skyux/packages/polyfills"
+              "zone.js/testing"
             ],
             "tsConfig": "tsconfig.spec.json",
             "karmaConfig": "karma.conf.js",