+++ libs/components/docs-tools/src/lib/modules/showcase/code-example-types/code-example-types.pipe.ts
@@ -8,9 +8,7 @@ import { SKY_DOCS_CODE_EXAMPLE_COMPONENT_TYPES } from './code-example-types-toke
 @Pipe({
   name: 'skyDocsCodeExampleNameToComponentType',
 })
-export class SkyDocsCodeExampleNameToComponentTypePipe
-  implements PipeTransform
-{
+export class SkyDocsCodeExampleNameToComponentTypePipe implements PipeTransform {
   readonly #examples = inject(SKY_DOCS_CODE_EXAMPLE_COMPONENT_TYPES);
 
   public transform(componentName: string): Type<unknown> {