+++ libs/components/manifest/src/types/function-def.ts
@@ -4,8 +4,7 @@ import type { SkyManifestParentDefinition } from './base-def';
  * Information about a function exported from the public API.
  * @internal
  */
-export interface SkyManifestFunctionDefinition
-  extends SkyManifestParentDefinition {
+export interface SkyManifestFunctionDefinition extends SkyManifestParentDefinition {
   kind: 'function';
   parameters?: SkyManifestParameterDefinition[];
   type: string;