+++ libs/sdk/e2e-schematics/src/generators/component-e2e/index.spec.ts
@@ -54,16 +54,6 @@ describe('component-e2e', () => {
     for (const projectName of ['test-storybook', 'test-storybook-e2e']) {
       config[projectName] = readProjectConfiguration(tree, projectName);
       expect(config[projectName].projectType).toEqual('application');
-      if (projectName === 'test-storybook') {
-        expect(
-          config[projectName].targets?.['build'].options.polyfills,
-        ).toBeDefined();
-        expect(
-          config[projectName].targets?.['build'].options.polyfills.includes(
-            'libs/components/packages/src/polyfills.js',
-          ),
-        ).toBeTruthy();
-      }
     }
     expect(
       tree.exists(`${config['test-storybook'].root}/.storybook/main.ts`),