+++ CHANGELOG.md
@@ -1,6 +1,80 @@
 # Changelog
 
 
+## [13.14.3](https://github.com/blackbaud/skyux/compare/13.14.2...13.14.3) (2026-03-03)
+
+
+### Bug Fixes
+
+* update to `@angular/build@20.3.18` ([#4255](https://github.com/blackbaud/skyux/issues/4255)) ([f822ac9](https://github.com/blackbaud/skyux/commit/f822ac986316791797bb1a5019bcd1d967c2eb96))
+
+
+
+## [14.0.0-alpha.5](https://github.com/blackbaud/skyux/compare/14.0.0-alpha.4...14.0.0-alpha.5) (2026-03-05)
+
+
+### ⚠ BREAKING CHANGES
+
+* **sdk/skyux-eslint:** add no-barrel-exports rule (#4268)
+
+### Features
+
+* **components/ag-grid:** use datepicker component for column filter ([#4249](https://github.com/blackbaud/skyux/issues/4249)) ([d026f28](https://github.com/blackbaud/skyux/commit/d026f280dd4b12cae271e7836c431f99ac1a37bf)), closes [AB#3648062](https://dev.azure.com/blackbaud/Products/_workitems/edit/3648062)
+* **sdk/skyux-eslint:** add no-barrel-exports rule ([#4268](https://github.com/blackbaud/skyux/issues/4268)) ([d7a8332](https://github.com/blackbaud/skyux/commit/d7a833296bd7b856e91df7f331f3454633650627))
+
+
+### Bug Fixes
+
+* update to `@angular/build@21.2.0` ([#4266](https://github.com/blackbaud/skyux/issues/4266)) ([e2c8ec0](https://github.com/blackbaud/skyux/commit/e2c8ec0121ad403b8228b7212600a916541dd41f))
+
+## [14.0.0-alpha.4](https://github.com/blackbaud/skyux/compare/14.0.0-alpha.3...14.0.0-alpha.4) (2026-03-03)
+
+
+### ⚠ BREAKING CHANGES
+
+* **sdk/skyux-eslint:** add prefer-input-box rule (#4243)
+* **sdk/skyux-eslint:** modify `skyux-eslint-template/prefer-label-text` ESLint rule to require `labelText` for all `sky-input-box` implementations (#4245)
+* **components/modals:** deprecate `SkyModalHeaderComponent` (#4244)
+
+### Features
+
+* **components/modals:** deprecate `SkyModalHeaderComponent` ([#4244](https://github.com/blackbaud/skyux/issues/4244)) ([5c80f39](https://github.com/blackbaud/skyux/commit/5c80f39f7ddc02bfd9fdd0fcdbabe145eeda5b75))
+* **sdk/skyux-eslint:** add prefer-input-box rule ([#4243](https://github.com/blackbaud/skyux/issues/4243)) ([bed5acf](https://github.com/blackbaud/skyux/commit/bed5acf45f883aac74529510f22860097a319a4a))
+
+
+### Bug Fixes
+
+* **sdk/skyux-eslint:** modify `skyux-eslint-template/prefer-label-text` ESLint rule to require `labelText` for all `sky-input-box` implementations ([#4245](https://github.com/blackbaud/skyux/issues/4245)) ([842e235](https://github.com/blackbaud/skyux/commit/842e235d6de47070a67889c9abb801ac21d54638))
+
+## [13.14.2](https://github.com/blackbaud/skyux/compare/13.14.1...13.14.2) (2026-02-25)
+
+
+### Bug Fixes
+
+* **components/list-builder:** stop overriding dropdown button padding ([#4181](https://github.com/blackbaud/skyux/issues/4181)) ([ca99b56](https://github.com/blackbaud/skyux/commit/ca99b56c2bcc3f208fbaec31cd49c3bb30067216))
+* **components/lookup:** lookup and selections modals use correct footer button spacing in modern theme ([#4229](https://github.com/blackbaud/skyux/issues/4229)) ([6453f81](https://github.com/blackbaud/skyux/commit/6453f81286687035440e68106dfc9f2762c2dc7e)), closes [AB#3704526](https://github.com/AB/issues/3704526)
+* use accessible text colors via design tokens update ([#4234](https://github.com/blackbaud/skyux/issues/4234)) ([ade98fb](https://github.com/blackbaud/skyux/commit/ade98fbf5cbc97bb103b8aa17daf51ee44d7c2f0))
+
+
+
+## [14.0.0-alpha.3](https://github.com/blackbaud/skyux/compare/14.0.0-alpha.2...14.0.0-alpha.3) (2026-02-25)
+
+
+### ⚠ BREAKING CHANGES
+
+* **components/packages:** do not add `@skyux/packages/polyfills` to new SKY UX projects (#4242)
+* **components/ag-grid:** support AG Grid 35.1.0 (#4221)
+
+### Features
+
+* **components/ag-grid:** support AG Grid 35.1.0 ([#4221](https://github.com/blackbaud/skyux/issues/4221)) ([41cd068](https://github.com/blackbaud/skyux/commit/41cd068af7ae82ce682a008445a80765d1417a4e))
+* **components/packages:** do not add `@skyux/packages/polyfills` to new SKY UX projects ([#4242](https://github.com/blackbaud/skyux/issues/4242)) ([8100400](https://github.com/blackbaud/skyux/commit/8100400d79c94a85772aec079313f91025e67011))
+
+
+### Bug Fixes
+
+* **components/packages:** install `ng2-dragula` if used in source code ([#4240](https://github.com/blackbaud/skyux/issues/4240)) ([f0ed850](https://github.com/blackbaud/skyux/commit/f0ed850c94caee254c2cb743c34a53dff4e6fffc))
+
 ## [14.0.0-alpha.2](https://github.com/blackbaud/skyux/compare/14.0.0-alpha.1...14.0.0-alpha.2) (2026-02-24)